Add method descriptions and improve project structure
A description was added to each method: this description is printed when selecting a given method. The project's structure was also improved: main.py stays in the root folder, but all other modules are moved to a "methods" folder, including a new utilities.py that contains portada() and clear_screen() and input_types.py, which was in splines/ before.